Add RSpec tests for models and update user validations
This commit introduces RSpec tests for Post, Comment, and User models. It tests the models for both associations and validations. Additionally, it ensures the User's password has a minimum length of 6 characters. Besides, bcrypt and shoulda-matchers gems have also been added to the Gemfile for password hashing and testing Rails models respectively.